Title: 바이러스 검사, Time: 352ms, Memory: 4MB, Status: Passed - CodeTree

+++ 250316/바이러스 검사/virus-detector.cpp
@@ -5,7 +5,7 @@ const int MAX_N = 1000000;
 
 int N;
 int arr[MAX_N];
-int result = 0;
+long long result = 0;
 int Leader, Follower;
 
 int main() {
